cleanup types renaming
